# tests/storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mik API",
        "description": "\u7c73\u5ba2\u4e4b\u5bb6API",
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/scans/logs": {
            "get": {
                "tags": [
                    "Scan Log"
                ],
                "summary": "\u7ec8\u7aef\u7684\u6838\u9500\u8bb0\u5f55",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "\u9875\u7801",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "\u6bcf\u9875\u663e\u793a\u6570\u91cf",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "301": {
                        "description": "\u5e10\u53f7\u672a\u901a\u8fc7\u5ba1\u6838\u4e0d\u80fd\u67e5\u770b"
                    },
                    "400": {
                        "description": "\u670d\u52a1\u5458\u6ca1\u6709\u6240\u5c5e\u7ec8\u7aef"
                    }
                }
            }
        },
        "/scans": {
            "post": {
                "tags": [
                    "Scan Log"
                ],
                "summary": "\u4e0a\u4f20\u626b\u63cf\u4e8c\u7ef4\u7801\u4fe1\u606f",
                "parameters": [
                    {
                        "name": "code",
                        "in": "formData",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "version",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u4e0a\u4f20\u626b\u63cf\u4e8c\u7ef4\u7801\u4fe1\u606f\u6210\u529f"
                    },
                    "400": {
                        "description": "\u5ba2\u6237\u7aef\u8bf7\u6c42\u6709\u8bef"
                    },
                    "500": {
                        "description": "\u670d\u52a1\u5668\u5185\u90e8\u9519\u8bef"
                    }
                }
            }
        },
        "/users/{user}/scans": {
            "get": {
                "tags": [
                    "Scan Log"
                ],
                "summary": "\u83b7\u53d6\u670d\u52a1\u5458\u6240\u6709\u626b\u7801\u8bb0\u5f55",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u626b\u7801\u8bb0\u5f55\u83b7\u53d6\u6210\u529f"
                    },
                    "500": {
                        "description": "\u670d\u52a1\u5668\u5185\u90e8\u9519\u8bef"
                    }
                }
            }
        },
        "/scans/rank": {
            "get": {
                "tags": [
                    "Scan Log"
                ],
                "summary": "\u6bd4\u4e00\u6bd4",
                "parameters": [
                    {
                        "name": "datefrom",
                        "in": "query",
                        "description": "e.g. 2015-12-10",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "dateto",
                        "in": "query",
                        "description": "e.g. 2016-12-10",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "areaid",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u6bd4\u4e00\u6bd4\u8bb0\u5f55\u83b7\u53d6\u6210\u529f"
                    }
                }
            }
        },
...
    },
...
}